requirements.txt
asgiref==3.11.1
br-cpf-cnpj==1.1.0
dj-database-url==3.1.2
Django==6.0.3
django-filter==25.2
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
Faker==40.13.0
gunicorn==25.3.0
inflection==0.5.1
Markdown==3.10.2
packaging==26.0
phonenumbers==9.0.27
psycopg2-binary==2.9.11
PyJWT==2.12.1
python-decouple==3.8
pytz==2026.1.post1
PyYAML==6.0.3
sqlparse==0.5.5
tzdata==2025.3
uritemplate==4.2.0
whitenoise==6.12.0
